requirements.txt
Django==1.10
django-filter==0.15.0
djangorestframework==3.4.6
selenium==2.53.6
